// src/Frontend/ViteApp/src/api-generated/models/PrintType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * KISv4 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PrintType = {
    DontPrint: 'DontPrint',
    PrintForCustomer: 'PrintForCustomer',
    PrintForEmployee: 'PrintForEmployee',
    PrintForBoth: 'PrintForBoth'
} as const;
export type PrintType = typeof PrintType[keyof typeof PrintType];


export function instanceOfPrintType(value: any): boolean {
    for (const key in PrintType) {
        if (Object.prototype.hasOwnProperty.call(PrintType, key)) {
            if (PrintType[key as keyof typeof PrintType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PrintTypeFromJSON(json: any): PrintType {
    return PrintTypeFromJSONTyped(json, false);
}

export function PrintTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PrintType {
    return json as PrintType;
}

export function PrintTypeToJSON(value?: PrintType | null): any {
    return value as any;
}

export function PrintTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): PrintType {
    return value as PrintType;
}