// src/Frontend/ViteApp/src/api-generated/models/SaleItemReadResponse.ts
export interface SaleItemReadResponse {
    /**
     * 
     * @type {CashBoxesReadIdParameter}
     * @memberof SaleItemReadResponse
     */
    id: CashBoxesReadIdParameter;
    /**
     * 
     * @type {string}
     * @memberof SaleItemReadResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SaleItemReadResponse
     */
    image: string | null;
    /**
     * 
     * @type {AccountTransactionModelAmount}
     * @memberof SaleItemReadResponse
     */
    marginPercent: AccountTransactionModelAmount;
    /**
     * 
     * @type {AccountTransactionModelAmount}
     * @memberof SaleItemReadResponse
     */
    marginStatic: AccountTransactionModelAmount;
    /**
     * 
     * @type {AccountTransactionModelAmount}
     * @memberof SaleItemReadResponse
     */
    prestigeAmount: AccountTransactionModelAmount;
    /**
     * 
     * @type {PrintType}
     * @memberof SaleItemReadResponse
     */
    printType: PrintType;
    /**
     * 
     * @type {Array<ModifierListModel>}
     * @memberof SaleItemReadResponse
     */
    applicableModifiers: Array<ModifierListModel>;
    /**
     * 
     * @type {Array<CategoryModel>}
     * @memberof SaleItemReadResponse
     */
    categories: Array<CategoryModel>;
}