// src/Frontend/ViteApp/src/api-generated/models/SaleItemUpdateModel.ts
export interface SaleItemUpdateModel {
    /**
     * 
     * @type {string}
     * @memberof SaleItemUpdateModel
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof SaleItemUpdateModel
     */
    image?: string | null;
    /**
     * 
     * @type {ModifierCreateRequestMarginPercent}
     * @memberof SaleItemUpdateModel
     */
    marginPercent?: ModifierCreateRequestMarginPercent;
    /**
     * 
     * @type {ModifierCreateRequestMarginStatic}
     * @memberof SaleItemUpdateModel
     */
    marginStatic?: ModifierCreateRequestMarginStatic;
    /**
     * 
     * @type {ModifierCreateRequestMarginStatic}
     * @memberof SaleItemUpdateModel
     */
    prestigeAmount?: ModifierCreateRequestMarginStatic;
    /**
     * 
     * @type {PrintType}
     * @memberof SaleItemUpdateModel
     */
    printType?: PrintType;
    /**
     * 
     * @type {Array<CashBoxesReadIdParameter>}
     * @memberof SaleItemUpdateModel
     */
    categoryIds?: Array<CashBoxesReadIdParameter>;
    /**
     * 
     * @type {Array<CashBoxesReadIdParameter>}
     * @memberof SaleItemUpdateModel
     */
    modifierIds?: Array<CashBoxesReadIdParameter>;
}